fixbug log format incorrect

+++ main/src/debug.c
@@ -80,7 +80,7 @@ void debug_thread(void* par) {
 	while (1) {
 		ulTaskNotifyTake(pdTRUE, ms_to_ticks(1000));
 		ESP_LOGI(TAG, "Debug thread running");
-		ESP_LOGI(TAG, "Available slots: %d", debug_get_available_slots());
+		ESP_LOGI(TAG, "Available slots: %d", (int)debug_get_available_slots());
 		if (debug_buf_head != debug_buf_end) {
 			xSemaphoreTake(debug_buf_mux, portMAX_DELAY);
